LSEEK IS FUCKED UP

+++ kern/syscall/file_syscall.c
@@ -60,7 +60,7 @@ off_t sys_lseek(int fd,off_t pos, int whence, int *retval){
 
 	struct filehandler *fh = curproc->filetable[fd];
 
-
+	
 
 	if(fh==NULL){
 		*retval=EBADF;
@@ -83,7 +83,7 @@ off_t sys_lseek(int fd,off_t pos, int whence, int *retval){
 			return -1;
 		}
 		kprintf("curoff:%dpos:%dnew:%d\n",fh->offset,(int)pos,(fh->offset+(int)pos));	
-		fh->offset=fh->offset+pos;	
+		fh->offset=fh->offset+(int)pos;	
 		*retval=fh->offset;
 		return 0;
